+++ msgfmt:core/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name:    "msgfmt:core",
-  version: "2.0.0-preview.13",
+  version: "2.0.0-preview.14",
   summary: "MessageFormat support, the Meteor way",
   git:     "https://github.com/gadicc/meteor-messageformat.git",
 });
@@ -9,6 +9,10 @@ Package.registerBuildPlugin({
   name: 'msgfmt',
   sources: [ 'buildPlugin.js' ]
 });
+
+var client = 'client';
+var server = 'server';
+var both = [ client, server ];
 
 Package.onUse(function (api) {
   api.versionsFrom("METEOR@1.0.1");
@@ -21,28 +25,37 @@ Package.onUse(function (api) {
     'meteorhacks:inject-initial@1.0.2',
     'jag:pince@0.0.6',
     'raix:eventemitter@0.1.2'
-  ], [ 'client', 'server' ]);
+  ], both);
 
   // client deps
   api.use([
     // core MDG packages
-    'templating', 'session', 'tracker', 'reactive-var', 'ddp',
+    'templating',
+    'session',
+    'tracker',
+    'reactive-var',
+    'ddp',
     // core 3rd party packages
-    'jquery'
-  ], 'client');
+    'jquery',
+    // MDG maintained non-core
+    'amplify@1.0.0'
+  ], client);
 
   // For msgfmt.storeUserLocale == true.
-  api.use('accounts-base', 'client', { weak: true });
+  api.use('accounts-base', client, { weak: true });
+
+  // For v0 warning check
+  api.use('gadicohen:messageformat@0.0.1', server, { weak: true });
 
   // api.use('browser-policy', 'server', { /* weak: true */ });
-
-  // MDG maintained non-core
-  api.use('amplify@1.0.0', 'client');
 
   //api.use(['ui', 'spacebars', 'htmljs'], 'client');
 
   // server deps
-  api.use(['webapp'], 'server');
+  api.use([
+    'check',
+    'webapp'
+  ], server);
 
   // common adds
 //  api.addFiles('lib/messageformat.js/messageformat.js', ['client', 'server']);
@@ -50,25 +63,25 @@ Package.onUse(function (api) {
   // TODO, locales by demand, respect namespacing by wrapping files ourselves
   api.addFiles([
     'lib/intl-messageformat/dist/intl-messageformat-with-locales.js'
-  ], [ 'client', 'server' ])
+  ], both)
 
   api.addFiles([
     'lib/mfPkg/messageformat.js'//,
 //    'lib/mfPkg/locale-all.js'
-  ], ['client', 'server']);
+  ], both);
 
   // server
-  api.addFiles('lib/mfPkg/messageformat-server.js', 'server');
+  api.addFiles('lib/mfPkg/messageformat-server.js', server);
 
   // client
   api.addFiles([
     'lib/mfPkg/messageformat.html',
     'lib/mfPkg/messageformat-client.js'
-  ], 'client');
+  ], client);
 
   // TODO, on cordova, need to bundle all languages
 
-  api.export(['mfPkg', 'mf', 'msgfmt'], ['client', 'server']);
+  api.export(['mfPkg', 'mf', 'msgfmt'], both);
 });
 
 Package.onTest(function(api) {
@@ -76,8 +89,8 @@ Package.onTest(function(api) {
   api.use('msgfmt:core');
   api.use('underscore');
   api.use('http', 'server');
-  api.use('browser-policy', 'server');
+  api.use('browser-policy', server);
 
-  api.addFiles('tests/tests-client.js', 'client');
-  api.addFiles('tests/tests-server.js', 'server');
+  api.addFiles('tests/tests-client.js', client);
+  api.addFiles('tests/tests-server.js', server);
 });